Merged PR 1589667: Thumbnail index fix
Correcting the CreateTexture function to pass the requested texture index/id to LoadImageBuffer instead of the image index/id.  This happened to work before because we normally have a 1-to-1 relationship between textures and images, but when we don't this breaks the cache loading math.

Bug 16500028 - [PC App] Most assets imported from Reinwood don't have thumbnails inside Woodinville's asset library