[#D3D-4573] Fixes download api calls
A call to /download was made when selecting a model
in order to get the size of its archive.
It was wrong since these calls were incrementing the download count
of the model.

Archive size has been moved to search results, so now /download
is called when the user actually download the asset.