Improve the performance of skinned animation import
Current implementation of importing animation curve is very slow for animations with lots of key frames. (e.g. motion capture animation ).
The bottleneck is in AnimationCurve.keys return a copy instead of reference. Hence, updating key frame in AnimationCurve is expensive. I build a list of keyframe for the curve, calculate the tangents from interpolation mode and then feed the key frames to the curve to avoid the overhead.