Improve the performance of skinned animation import (#360)
* Improve the performance of skinned animation import

Current implementation of importing animation curve is very slow for animations with lots of key frames. (e.g. motion capture animation ).
The bottleneck is in AnimationCurve.keys return a copy instead of reference. Hence, updating key frame in AnimationCurve is expensive. I build a list of keyframe for the curve, calculate the tangents from interpolation mode and then feed the key frames to the curve to avoid the overhead.

* Improve the performance of skinned animation import

1, put the delegate function at the top of the file
2, replace spaces with tab
3, replace the keyframe list cache with 2d array