Better handling of attribute errors
Instead of throwing when finding unrecognized attributes, this PR makes sure all attributes are collected and an exception is thrown after all are known. This allows to actually catch the exception and handle it per object instead of just failing.
Note that I did this because TEXCOORD_4 throws exceptions (it's not available in the attribute list) - not sure if this is a bug in itself.