Fix emission color export for re-exported GLTF files
When re-exporting a mesh that was imported via GLTF, the emission part of the material wasn't being exported correctly due to the order of emission properties being checked. This commit changes the order of properties checked to ensure `emissiveFactor` is checked first, followed by `_EmissionColor`, and then `_EmissiveFactor`, thus correctly handling the case of re-exporting GLTF files.